+++ CHANGELOG.md
@@ -17,3 +17,4 @@ For information on changes in released versions of Teku, see the [releases page]
 - Added support for exporting metrics to an external consumer with `--metrics-publish-endpoint`.
 
 ### Bug Fixes
+ - Improved handling of the http accept header used to determine whether to send SSZ data or json for states and blocks.